requirements.txt
langchain
langchain-community
faiss-cpu
sentence-transformers
google-generativeai
pypdf
fastapi
uvicorn
python-multipart
tqdm
aiofiles  # for async file uploads
jinja2     # Optional, for templating
langchainhub  # For community tools/templates
langchain-cli # For dev tools and chains
requests
torch